Requisito 09 - Atualize a quantidade de produtos. :white_check_mark:

+++ models/salesModel.js
@@ -1,7 +1,18 @@
 const connection = require('./connections');
 const { ObjectId } = require('mongodb');
 
+const { getByIdProduct, updateProduct } = require('./productsModel');
+
 const createSales = async (newSales) => {
+  const zero = 0;
+  const { productId, quantity } = newSales[0];
+
+  const product = await getByIdProduct(productId);
+  if (!product) return null;
+  const newQuantity = product.quantity - quantity;
+  if (newQuantity < zero) return null;
+  await updateProduct(productId, product.name, newQuantity);
+
   const result = await connection()
     .then((db) => db
       .collection('sales')
@@ -40,10 +51,20 @@ const updateSale = async (id, updates) => {
 
 const deleteSale = async (id) => {
   if (!ObjectId.isValid(id)) return null;
+
+  const sales = await getByIdSale(id);
+  if (!sales) return null;
+  const { productId, quantity: quantitySales } = sales.itensSold[0];
+
+  const { name, quantity: quantityProduct } = await getByIdProduct(productId);
+  const newQuantity = quantityProduct + quantitySales;
+  await updateProduct(productId, name, newQuantity);
+
   const result = await connection()
     .then((db) => db
       .collection('sales')
       .findOneAndDelete({ _id: ObjectId(id) }));
+
   return result.value;
 };
